+++ include/openthread.h
@@ -842,7 +842,7 @@ uint32_t otGetNetworkIdTimeout(void);
  *
  * @sa otGetNetworkIdTimeout
  */
-void otSetNetworkIdTimeout(uint32_t aTimeout);
+void otSetNetworkIdTimeout(uint8_t aTimeout);
 
 /**
  * Get the ROUTER_UPGRADE_THRESHOLD parameter used in the REED role.
