+++ src/ncp/ncp_base.cpp
@@ -47,7 +47,7 @@ extern ThreadNetif *sThreadNetif;
 
 static NcpBase *sNcpContext = NULL;
 
-#define NCP_PLAT_RESET_REASON        (1<<31)
+#define NCP_PLAT_RESET_REASON        (1U<<31)
 
 // ----------------------------------------------------------------------------
 // MARK: Command/Property Jump Tables
@@ -389,7 +389,7 @@ void NcpBase::HandleActiveScanResult(otActiveScanResult *result)
 {
     if (result)
     {
-        uint8_t flags = (result->mVersion << SPINEL_BEACON_THREAD_FLAG_VERSION_SHIFT);
+        uint8_t flags = static_cast<uint8_t>(result->mVersion << SPINEL_BEACON_THREAD_FLAG_VERSION_SHIFT);
 
         if (result->mIsJoinable)
         {
@@ -458,7 +458,7 @@ void NcpBase::UpdateChangedProps(void)
     {
         if ((mChangedFlags & NCP_PLAT_RESET_REASON) != 0)
         {
-            mChangedFlags &= ~NCP_PLAT_RESET_REASON;
+            mChangedFlags &= ~static_cast<uint32_t>(NCP_PLAT_RESET_REASON);
             SendLastStatus(
                 SPINEL_HEADER_FLAG | SPINEL_HEADER_IID_0,
                 ResetReasonToSpinelStatus(otPlatGetResetReason())
@@ -466,7 +466,7 @@ void NcpBase::UpdateChangedProps(void)
         }
         else if ((mChangedFlags & OT_IP6_LL_ADDR_CHANGED) != 0)
         {
-            mChangedFlags &= ~OT_IP6_LL_ADDR_CHANGED;
+            mChangedFlags &= ~static_cast<uint32_t>(OT_IP6_LL_ADDR_CHANGED);
             HandleCommandPropertyGet(
                 SPINEL_HEADER_FLAG | SPINEL_HEADER_IID_0,
                 SPINEL_PROP_IPV6_LL_ADDR
@@ -478,7 +478,7 @@ void NcpBase::UpdateChangedProps(void)
                               SPINEL_HEADER_FLAG | SPINEL_HEADER_IID_0,
                               SPINEL_PROP_IPV6_ML_ADDR
                           ));
-            mChangedFlags &= ~OT_IP6_ML_ADDR_CHANGED;
+            mChangedFlags &= ~static_cast<uint32_t>(OT_IP6_ML_ADDR_CHANGED);
         }
         else if ((mChangedFlags & OT_NET_STATE) != 0)
         {
@@ -486,7 +486,7 @@ void NcpBase::UpdateChangedProps(void)
                               SPINEL_HEADER_FLAG | SPINEL_HEADER_IID_0,
                               SPINEL_PROP_NET_STATE
                           ));
-            mChangedFlags &= ~OT_NET_STATE;
+            mChangedFlags &= ~static_cast<uint32_t>(OT_NET_STATE);
         }
         else if ((mChangedFlags & OT_NET_ROLE) != 0)
         {
@@ -494,7 +494,7 @@ void NcpBase::UpdateChangedProps(void)
                               SPINEL_HEADER_FLAG | SPINEL_HEADER_IID_0,
                               SPINEL_PROP_NET_ROLE
                           ));
-            mChangedFlags &= ~OT_NET_ROLE;
+            mChangedFlags &= ~static_cast<uint32_t>(OT_NET_ROLE);
 
         }
         else if ((mChangedFlags & OT_NET_PARTITION_ID) != 0)
@@ -503,7 +503,7 @@ void NcpBase::UpdateChangedProps(void)
                               SPINEL_HEADER_FLAG | SPINEL_HEADER_IID_0,
                               SPINEL_PROP_NET_PARTITION_ID
                           ));
-            mChangedFlags &= ~OT_NET_PARTITION_ID;
+            mChangedFlags &= ~static_cast<uint32_t>(OT_NET_PARTITION_ID);
         }
         else if ((mChangedFlags & OT_NET_KEY_SEQUENCE) != 0)
         {
@@ -511,7 +511,7 @@ void NcpBase::UpdateChangedProps(void)
                               SPINEL_HEADER_FLAG | SPINEL_HEADER_IID_0,
                               SPINEL_PROP_NET_KEY_SEQUENCE
                           ));
-            mChangedFlags &= ~OT_NET_KEY_SEQUENCE;
+            mChangedFlags &= ~static_cast<uint32_t>(OT_NET_KEY_SEQUENCE);
         }
         else if ((mChangedFlags & (OT_IP6_ADDRESS_ADDED | OT_IP6_ADDRESS_REMOVED)) != 0)
         {
@@ -519,7 +519,7 @@ void NcpBase::UpdateChangedProps(void)
                               SPINEL_HEADER_FLAG | SPINEL_HEADER_IID_0,
                               SPINEL_PROP_IPV6_ADDRESS_TABLE
                           ));
-            mChangedFlags &= ~(OT_IP6_ADDRESS_ADDED | OT_IP6_ADDRESS_REMOVED);
+            mChangedFlags &= ~static_cast<uint32_t>(OT_IP6_ADDRESS_ADDED | OT_IP6_ADDRESS_REMOVED);
         }
         else if ((mChangedFlags & (OT_THREAD_CHILD_ADDED | OT_THREAD_CHILD_REMOVED)) != 0)
         {
@@ -528,7 +528,7 @@ void NcpBase::UpdateChangedProps(void)
             //    SPINEL_HEADER_FLAG | SPINEL_HEADER_IID_0,
             //    SPINEL_PROP_THREAD_CHILD_TABLE
             //));
-            mChangedFlags &= ~(OT_THREAD_CHILD_ADDED | OT_THREAD_CHILD_REMOVED);
+            mChangedFlags &= ~static_cast<uint32_t>(OT_THREAD_CHILD_ADDED | OT_THREAD_CHILD_REMOVED);
         }
     }
 
@@ -853,7 +853,7 @@ ThreadError NcpBase::OutboundFrameFeedVPacked(const char *pack_format, va_list a
 
     if ((packed_len > 0) && (packed_len <= static_cast<spinel_ssize_t>(sizeof(buf))))
     {
-        errorCode = OutboundFrameFeedData(buf, packed_len);
+        errorCode = OutboundFrameFeedData(buf, static_cast<uint16_t>(packed_len));
     }
 
     return errorCode;
@@ -944,7 +944,8 @@ ThreadError NcpBase::CommandHandler_PROP_VALUE_SET(uint8_t header, unsigned int 
 
     if (parsedLength == arg_len)
     {
-        errorCode = HandleCommandPropertySet(header, static_cast<spinel_prop_key_t>(propKey), value_ptr, value_len);
+        errorCode = HandleCommandPropertySet(header, static_cast<spinel_prop_key_t>(propKey), value_ptr,
+					     static_cast<uint16_t>(value_len));
     }
     else
     {
@@ -969,7 +970,8 @@ ThreadError NcpBase::CommandHandler_PROP_VALUE_INSERT(uint8_t header, unsigned i
 
     if (parsedLength == arg_len)
     {
-        errorCode = HandleCommandPropertyInsert(header, static_cast<spinel_prop_key_t>(propKey), value_ptr, value_len);
+        errorCode = HandleCommandPropertyInsert(header, static_cast<spinel_prop_key_t>(propKey), value_ptr,
+						static_cast<uint16_t>(value_len));
     }
     else
     {
@@ -994,7 +996,8 @@ ThreadError NcpBase::CommandHandler_PROP_VALUE_REMOVE(uint8_t header, unsigned i
 
     if (parsedLength == arg_len)
     {
-        errorCode = HandleCommandPropertyRemove(header, static_cast<spinel_prop_key_t>(propKey), value_ptr, value_len);
+        errorCode = HandleCommandPropertyRemove(header, static_cast<spinel_prop_key_t>(propKey), value_ptr,
+						static_cast<uint16_t>(value_len));
     }
     else
     {
@@ -2555,7 +2558,7 @@ ThreadError NcpBase::SetPropertyHandler_STREAM_NET_INSECURE(uint8_t header, spin
         (void)meta_len;
         (void)parsedLength;
 
-        errorCode = message->Append(frame_ptr, frame_len);
+        errorCode = message->Append(frame_ptr, static_cast<uint16_t>(frame_len));
     }
 
     if (errorCode == kThreadError_None)
@@ -2623,7 +2626,7 @@ ThreadError NcpBase::SetPropertyHandler_STREAM_NET(uint8_t header, spinel_prop_k
         (void)meta_len;
         (void)parsedLength;
 
-        errorCode = message->Append(frame_ptr, frame_len);
+        errorCode = message->Append(frame_ptr, static_cast<uint16_t>(frame_len));
     }
 
     if (errorCode == kThreadError_None)
@@ -3342,6 +3345,6 @@ ThreadError otNcpStreamWrite(int aStreamId, const uint8_t* aDataPtr, int aDataLe
         SPINEL_CMD_PROP_VALUE_IS,
         static_cast<spinel_prop_key_t>(aStreamId),
         aDataPtr,
-        aDataLen
+        static_cast<uint16_t>(aDataLen)
     );
 }
