+++ src/ncp/ncp_buffer.cpp
@@ -131,12 +131,12 @@ uint16_t NcpFrameBuffer::GetDistance(uint8_t *aStartPtr, uint8_t *aEndPtr) const
 
     if (aEndPtr >= aStartPtr)
     {
-        distance =  aEndPtr - aStartPtr;
+        distance = static_cast<size_t>(aEndPtr - aStartPtr);
     }
     else
     {
-        distance  = mBufferEnd - aStartPtr;
-        distance +=  aEndPtr - mBuffer;
+        distance  = static_cast<size_t>(mBufferEnd - aStartPtr);
+        distance += static_cast<size_t>(aEndPtr - mBuffer);
     }
 
     return static_cast<uint16_t>(distance);
@@ -154,7 +154,7 @@ uint16_t NcpFrameBuffer::ReadUint16At(uint8_t *aBufPtr)
 {
     uint16_t value;
 
-    value = (*aBufPtr) << 8;
+    value = static_cast<uint16_t>((*aBufPtr) << 8);
     value += *Next(aBufPtr);
 
     return value;
@@ -355,7 +355,8 @@ ThreadError NcpFrameBuffer::OutFramePrepareSegment(void)
         }
 
         // Find tail/end of current segment.
-        mReadSegmentTail = Advance(mReadSegmentHead, kSegmentHeaderSize + (header & kSegmentHeaderLengthMask));
+        mReadSegmentTail = Advance(mReadSegmentHead,
+				   kSegmentHeaderSize + static_cast<uint8_t>(header & kSegmentHeaderLengthMask));
 
         // Update the current read pointer to skip the segment header.
         mReadPointer = Advance(mReadSegmentHead, kSegmentHeaderSize);
@@ -576,7 +577,7 @@ ThreadError NcpFrameBuffer::OutFrameRemove(void)
         }
 
         // Move the pointer to next segment.
-        bufPtr = Advance(bufPtr, kSegmentHeaderSize + (header & kSegmentHeaderLengthMask));
+        bufPtr = Advance(bufPtr, kSegmentHeaderSize + static_cast<uint8_t>(header & kSegmentHeaderLengthMask));
     }
 
     mReadFrameStart = bufPtr;
@@ -644,7 +645,7 @@ uint16_t NcpFrameBuffer::OutFrameGetLength(void)
         frameLength += (header & kSegmentHeaderLengthMask);
 
         // Move the pointer to next segment.
-        bufPtr = Advance(bufPtr, kSegmentHeaderSize + (header & kSegmentHeaderLengthMask));
+        bufPtr = Advance(bufPtr, kSegmentHeaderSize + static_cast<uint8_t>(header & kSegmentHeaderLengthMask));
     }
 
     // Remember the calculated frame length for current frame.
